some checks

+++ note.sh
@@ -406,11 +406,12 @@ cmd_delete() {
 }
 
 cmd_rename() {
+    die_if_invalid_path "$1"
     die_if_invalid_path "$2"
     die_if_name_not_entered "$1"
     die_if_name_not_entered "$2"
-    test -e "$PREFIX/$1" || die "Note '$1' doesn\`t exist" $INVALID_ARG_CODE
-    test -f "$PREFIX/$2" && die "Note '$2' already exists" $INVALID_ARG_CODE
+    test -e "$PREFIX/$1" || die "Note or directory '$1' doesn\`t exist" $INVALID_ARG_CODE
+    test -f "$PREFIX/$2" && die "Note or directory '$2' already exists" $INVALID_ARG_CODE
 
     _DIRNAME="$(dirname "$2")"
 
@@ -425,12 +426,13 @@ cmd_rename() {
 }
 
 cmd_ln() {
+    die_if_invalid_path "$1"
     die_if_invalid_path "$2"
     die_if_name_not_entered "$1"
     die_if_name_not_entered "$2"
 
-    test -e "$PREFIX/$1" || die "Note '$1' doesn\`t exist" $INVALID_ARG_CODE
-    test -e "$PREFIX/$2" && die "Note '$2' already exists" $INVALID_ARG_CODE
+    test -e "$PREFIX/$1" || die "Note or directory '$1' doesn\`t exist" $INVALID_ARG_CODE
+    test -f "$PREFIX/$2" && die "Note or directory '$2' already exists" $INVALID_ARG_CODE
 
     ln -s "$PREFIX/$1" "$PREFIX/$2"
     git_add "$2"
